upadte

+++ function3.c
@@ -2,9 +2,9 @@
 #include <math.h> 
 int main() {
     printf("Enter a numbers whose square you want to get : ");
-    float a; 
-   scanf("%f", &a);
-    printf("The square of %f is %.2f\n", a , pow(a, 2));
+     int a; 
+   scanf("%d", &a);
+    printf("The square of %d is %.0f\n", a , pow(a, 2));
     return 0 ;
 }
